# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MrRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MrRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/WCY-dt/MrRSS",
            "email": "mail@ch3nyang.top"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.3.14"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
...
        "/articles/filter": {
            "post": {
                "description": "Retrieve articles with advanced filtering conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get filtered articles",
                "parameters": [
                    {
                        "description": "Filter criteria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/article.FilterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered articles",
                        "schema": {
                            "$ref": "#/definitions/article.FilterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/browser/open": {
            "post": {
                "description": "Returns redirect response for client-side URL opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "browser"
                ],
                "summary": "Open URL in browser (server mode)",
                "parameters": [
                    {
                        "description": "Open URL request (url)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect URL (redirect)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid URL)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "article.FilterCondition": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "\"feed_name\", \"feed_category\", \"article_title\", \"published_after\", \"published_before\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logic": {
                    "description": "\"and\", \"or\" (null for first condition)",
                    "type": "string"
                },
                "negate": {
                    "description": "NOT modifier for this condition",
                    "type": "boolean"
                },
                "operator": {
                    "description": "\"contains\", \"exact\" (null for date fields and multi-select)",
                    "type": "string"
                },
                "value": {
                    "description": "Single value for text/date fields",
                    "type": "string"
                },
                "values": {
                    "description": "Multiple values for feed_name and feed_category",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "article.FilterRequest": {
            "type": "object",
            "properties": {
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/article.FilterCondition"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "article.FilterResponse": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Article"
                    }
                },
                "has_more": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.Article": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "integer"
                },
                "feed_title": {
                    "description": "Joined field",
                    "type": "string"
                },
                "freshrss_item_id": {
                    "description": "FreshRSS/Google Reader item ID for API operations",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "is_read": {
                    "type": "boolean"
                },
                "is_read_later": {
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "summary": {
                    "description": "Cached AI-generated summary",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "translated_title": {
                    "type": "string"
                },
                "unique_id": {
                    "description": "Unique identifier for deduplication (title+feed_id+published_date)",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "video_url": {
                    "description": "YouTube video URL for embedded player",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}